requirements.txt
yfinance
textblob
pandas
requests
python-dotenv
alpha-vintage
plotly